Fix name to char

+++ src/game/troops/CCoord.h
@@ -2,7 +2,7 @@
 #include <string>
 
 struct CCoord {
-    CCoord ( int c, int r ) : m_Colum (c), m_Row(r) {};
+    CCoord ( size_t c, size_t r ) : m_Colum ( c), m_Row( r) {};
     CCoord () = default;
     /* coord is a sttring with human readible string */
     explicit CCoord ( const std::string & coord );
